backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
bcrypt==4.1.2
alibabacloud-credentials==1.0.3
alibabacloud-credentials-api==1.0.0
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.4.1
alibabacloud-tingwu20230930==2.0.24
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_util==0.3.13
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
crcmod==1.7
cryptography==44.0.3
darabonba-core==1.0.4
ecdsa==0.19.1
fastapi==0.121.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==0.10.0
librosa==0.10.2
loguru==0.7.3
multidict==6.7.0
numpy==1.26.4
oss2==2.19.1
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
yarl==1.22.0
